pyproject.toml
[tool.poetry]
name = "goodmandu"
version = "0.1.0"
description = ""
authors = ["keshab"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "^2.2"
django-rest-framework = "^0.1.0"
django-countries = "^5.5"
django-taggit = "^1.1"
gunicorn = "^19.9"
psycopg2 = "^2.8"
markdown2 = "^2.3"
Pillow = "^6.2"
django-webpack-loader = "^0.6.0"

[tool.poetry.dev-dependencies]
ipython = "^7.8"
django-debug-toolbar = "^2.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
